% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShotgunSeq.R
\name{shotgunfunction}
\alias{shotgunfunction}
\title{High dimensional statistics for shotgun metagenomic data}
\usage{
shotgunfunction(df, meta1, subjects, covariate, df_label,
  meta_label = "BioSample", ncores = 4, get_coef = get_coef_lme,
  get_pval = get_p_value_lme, q_cutoff = 0.2)
}
\arguments{
\item{df}{A dataframe or tibble.}

\item{meta1}{A dataframe or tibble with covariate info}

\item{subjects}{random effect}

\item{covariate}{fixed effect}

\item{df_label}{The column containing gene names in df}

\item{meta_label}{The column containing sample names in meta1}

\item{ncores}{Number of cores to use}

\item{get_coef}{Function to get effect size estimates}

\item{get_pval}{Function to get p-values}

\item{q_cutoff}{Cutoff for FDR control}
}
\description{
High dimensional statistics for shotgun metagenomic data
}
