\name{ShotgunSeq-package}
\alias{ShotgunSeq-package}
\alias{ShotgunSeq}
\docType{package}
\title{
\packageTitle{ShotgunSeq}
}
\description{
\packageDescription{ShotgunSeq}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ShotgunSeq}
\packageIndices{ShotgunSeq}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{ShotgunSeq}

Maintainer: \packageMaintainer{ShotgunSeq}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
